Add IdentifiedCollection
An IdentifiedCollection is a collection of items that can be identified
using a key. The IdentifiedCollection's constructor takes a delegate
that is used to derive the key used to look up an item. Items' keys are
generated by that KeyFinder when the item is inserted and the item may
be either looked up by this key or iterated over like a normal
collection.

The goals for this data structure is:

1. Provide an easy/relatively fast way to look up items in the
collection based on a key for an item.
2. Maintain that keys must be derivable by the items.
3. Maintain the order of the items' iteration as inserted (which is a
major distinction from a regular dictionary with manually managed keys).